# [DataFrame] Improve performance of iteration methods